Restructured fetching modules to models directory
