+++ src/fedora_easyfix/models/gitlab.py
@@ -22,34 +22,33 @@
 
 from json import loads
 
-from fedora_easyfix.utilities.compose import StatusDecorator
-from urllib3 import PoolManager, make_headers
+from fedora_easyfix.utilities.composer import StatusDecorator
+from urllib3 import PoolManager
 
 httpobjc = PoolManager()
-api_base_url = "https://api.github.com/repos/"
+api_base_url = "https://gitlab.com/api/v4/projects/"
 statdcrt = StatusDecorator()
 
 
-class GitHubRepositories():
-    def __init__(self, repository_list, base_url, api_key, username):
+class GitLabRepositories():
+    def __init__(self, repository_list, base_url, api_key):
         self.repository_list = repository_list
         self.base_url = base_url
         self.api_key = api_key
         self.repository_collection = {}
-        self.headers = make_headers(basic_auth="%s:%s" %(username, api_key))
 
     def fetch_tickets_from_repository(self, repository_name):
         label = self.repository_list[repository_name]["label"]
+        project_id = self.repository_list[repository_name]["pid"]
         contact = self.repository_list[repository_name]["contact"]
-        api_issue_endpoint = "%s%s/issues" % (api_base_url, repository_name)
+        api_issue_endpoint = "%s%s/issues" % (api_base_url, project_id)
         respobjc = httpobjc.request(
             "GET",
             api_issue_endpoint,
-            headers=self.headers,
             fields={
                 "per_page": 100,
                 "labels": label,
-                "state": "open"
+                "state": "opened"
             }
         )
         respdict = loads(respobjc.data)
@@ -57,36 +56,38 @@ class GitHubRepositories():
         ticket_list = {}
         for ticket in respdict:
             ticket_count += 1
-            ticket_list[ticket["number"]] = {
+            ticket_list[ticket["iid"]] = {
                 "title": ticket["title"],
                 "date_created": ticket["created_at"],
                 "last_updated": ticket["updated_at"],
                 "creator": {
-                    "full_url": ticket["user"]["html_url"],
-                    "name": ticket["user"]["login"]
+                    "full_url": ticket["author"]["web_url"],
+                    "fullname": ticket["author"]["name"],
+                    "name": ticket["author"]["username"]
                 },
-                "url": ticket["html_url"],
-                "labels": [label["name"] for label in ticket["labels"]]
+                "url": ticket["web_url"],
+                "labels": ticket["labels"]
             }
-        api_project_endpoint = "%s%s" % (api_base_url, repository_name)
+        api_project_endpoint = "%s%s" % (api_base_url, project_id)
         respobjc = httpobjc.request(
             "GET",
             api_project_endpoint,
-            headers=self.headers,
         )
         respdict = loads(respobjc.data)
         ticket_dict = {
             "ticket_count": ticket_count,
             "ticket_list": ticket_list,
             "contact": "%s@fedoraproject.org" % contact,
-            "url": respdict["html_url"],
+            "url": respdict["web_url"],
             "description": respdict["description"],
             "id": respdict["id"],
             "target_label": label,
             "maintainer": {
-                "full_url": respdict["owner"]["html_url"],
-                "name": respdict["owner"]["login"]
+                "full_url": respdict["namespace"]["web_url"],
+                "fullname": respdict["namespace"]["name"],
+                "name": respdict["namespace"]["path"]
             },
+            "tags": respdict["tag_list"],
             "date_created": respdict["created_at"]
         }
         return ticket_dict, ticket_count
